+++ src/main/java/com/back/together02be/trade/processor/TradeBuyProcessor.java
@@ -4,10 +4,10 @@ import com.back.together02be.asset.enitity.UserAccount;
 import com.back.together02be.asset.enitity.UserStock;
 import com.back.together02be.asset.repository.UserAccountRepository;
 import com.back.together02be.asset.repository.UserStockRepository;
-import com.back.together02be.stock.enitity.Stock;
+import com.back.together02be.stock.entity.Stock;
 import com.back.together02be.stock.repository.StockRepository;
 import com.back.together02be.stock.service.RealTimeStockPriceStore;
-import com.back.together02be.stock.service.RealtimeStockPrice;
+import com.back.together02be.stock.dto.RealtimeStockPrice;
 import com.back.together02be.trade.dto.BuyReq;
 import com.back.together02be.trade.dto.BuyRes;
 import com.back.together02be.trade.enitity.Trade;
