+++ src/test/java/com/back/together02be/trade/processor/TradeBuyProcessorTest.java
@@ -12,15 +12,15 @@ import com.back.together02be.asset.enitity.UserAccount;
 import com.back.together02be.asset.enitity.UserStock;
 import com.back.together02be.asset.repository.UserAccountRepository;
 import com.back.together02be.asset.repository.UserStockRepository;
-import com.back.together02be.stock.enitity.Stock;
+import com.back.together02be.stock.entity.Stock;
 import com.back.together02be.stock.repository.StockRepository;
 import com.back.together02be.stock.service.RealTimeStockPriceStore;
-import com.back.together02be.stock.service.RealtimeStockPrice;
+import com.back.together02be.stock.dto.RealtimeStockPrice;
 import com.back.together02be.trade.dto.BuyReq;
 import com.back.together02be.trade.dto.BuyRes;
 import com.back.together02be.trade.enitity.Trade;
 import com.back.together02be.trade.repository.TradeRepository;
-import com.back.together02be.users.enitity.Users;
+import com.back.together02be.users.entity.Users;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
